Add files via upload
Jupyter Notebook to pull data from FMP Cloud API to get financial information. I used pandas to organize the information and xlsxWriter to create Excel spreadsheets with charts. The information I focused on for this exercise was Revenue, Free Cash Flow, Earnings Per Share, and the percentage change of each.